Clean up code: * Seperate into pkg/tftp and cmds/core/tftp * Split tests accordingly * Add more tests
